requirements.txt
googletrans==4.0.2
translate==3.6.1
langdetect==1.0.9
pytubefix==8.12.1
opencv-python==4.8.1.78
pillow==10.2.0
face-recognition==1.3.0
deepface==0.0.92
numpy==1.26.4
nltk==3.8.1
spacy==3.8.2
pandas==2.2.2
matplotlib==3.10.1
seaborn==0.13.2
scipy==1.15.2
scikit-learn==1.5.2